requirements.txt
PyYAML>=3.12 # MIT
tosca-parser>=1.4.0 # Apache-2.0